test: Fix off-by-one in mockscheduler test RPC
Summary:
Make it possible to `MockForwardAndSync(1h)`

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/21948 | core#21948]]

Test Plan: The bug was found by Core devs via fuzzing.  I noticed it because core#33211 adds a `MockForwardAndSync(1h)`. So check that this makes the next commit pass tests without triggering the assertion.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20099