test: Use COINBASE_MATURITY in functional tests
Summary:
```
Goal of this PR is to replace integer constants with COINBASE_MATURITY but not necessarily in all cases because that would mean to read and fully understand all tests. That's out of my time constraints. Any reports where COINBASE_MATURITY should be used are welcome though!
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/21989 | core#21989]].

Test Plan:
With chronik:
  ninja check-functional-extended

Reviewers: #bitcoin_abc, roqqit

Reviewed By: roqqit

Subscribers: roqqit

Differential Revision: https://reviews.bitcoinabc.org/D16385