[secp256k1] ecmult_gen: Move table creation to new file and force static prec
Summary:
This is a partial backport of [[https://github.com/bitcoin-core/secp256k1/pull/988 | secp256k1#988]]
https://github.com/bitcoin-core/secp256k1/pull/988/commits/22dc2c0a0dc3b321e72253f492cfa8bcbf00169b

Note that this diff breaks the rust wrapping of the secp256k1 library and all the monorepo modules that depend on that rust wrapping. But everything is fixed again later in the same backport (D19400) when the precomputed table is included in the repo.

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19382