[secp256k1] Optimization: use Nx32 representation for recoded bits
Summary:
The existing code needs to deal with the edge case that bit_pos >= 256,
which would lead to an out-of-bounds read from secp256k1_scalar.

Instead, recode the scalar into an array of uint32_t with enough zero
padding at the end to alleviate the issue. This also simplifies the
code, and is necessary for a security improvement in a follow-up
commit.

Original code by Peter Dettman, with modifications by Pieter Wuille.

This is a partial backport of [[https://github.com/bitcoin-core/secp256k1/pull/1058 | secp256k1#1058]]
https://github.com/bitcoin-core/secp256k1/pull/1058/changes/a0d32b597d4f9c8a20547dd21677716d4df8da54

Depends on D20269

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20270