net: remove CService::ToString() use ToStringAddrPort() instead
Summary:
Both methods do the same thing, so simplify to having just one.

`ToString()` is too generic in this case and it is unclear what it does,
given that there are similar methods:
`ToStringAddr()` (inherited from `CNetAddr`),
`ToStringPort()` and
`ToStringAddrPort()`.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/25619 | core#25619]]
https://github.com/bitcoin/bitcoin/pull/25619/changes/96c791dd20fea54c17d224000dee677bc158f66a

Depends on D19690

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19691