rpc: warn that nodes ignore requests for old stale blocks
Summary:
Adds warning to RPC help that `getblockfrompeer` is of little use for stale blocks that are more than a month old.

This is an anti-fingerprinting measure. See `BlockRequestAllowed` in `net_processing`.

It's been in Bitcoin Core since 2014, introduced in #2910 and later improved to not rely on checkpoints.
Older and alternative clients might still serve these blocks, so not throwing an error.

Allowing whitelisted nodes to fetch these blocks anyway might be nice.

Backport of [[https://github.com/bitcoin/bitcoin/pull/24226 | core#24226]]

Reviewer note: This was an AI-automated backport.

Test Plan:
  ninja
  ./src/bitcoind -regtest
  ./src/bitcoin-cli -regtest help getblockfrompeer

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D20223