Use steady clock in SeedStrengthen and FlushStateToDisk
Summary:
There may be a theoretical deadlock for the duration of the offset when the system clock is adjusted into a past time while executing SeedStrengthen.

Fix this by using steady clock.

Do the same in FlushStateToDisk, which should make the flushes more steady, if the system clock is adjusted by a large offset.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/27189 | core#27189]]

Depends on D18636

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18638