refactor: Add stop_at_height option in ChainstateManager
Summary:
Remove access to the global gArgs for the stopatheight argument and
replace it by adding a field to the existing ChainstateManager Options
struct.

This should eventually allow users of the ChainstateManager to not rely
on the global gArgs and instead pass in their own options.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/27576 | core#27576]]
https://github.com/bitcoin/bitcoin/pull/27576/commits/ef95be334f3aec671346372b64606e0fd390979a

Depends on D18684

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18685