macOS: Bump minimum required runtime version and prepare for building with upstream LLVM
Summary:
```
[...]
with the minimum runtime version bumped to 11.0, our current version of pre-compiled clang we use for macOS is too old to understand -mmacosx-version-min=11.0 because it expects =10.x. So I've made the smallest possible bump (from 10.0.1 to 11.1.0) to a version that understands. This bump is arbitrary and unfortunate, but likely to be short-lived as we may end up replacing it with llvm+lld for v26 anyway. I've held off on bumping the SDK as I think that makes sense to do as part of the lld switch instead.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/27676 | core#27676]] and [[https://github.com/bitcoin/bitcoin/pull/27493 | core#27493]].

Also contains the source for an arm64 build of native_clang (untested).

Depends on D17208.

Test Plan: Run the OSX GUIX build.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17209