kernel: Add flushError method to notifications
Summary:
This is done in addition with the following commit. Both have the goal
of getting rid of direct calls to AbortNode from kernel code. This extra
flushError method is added to notify specifically about errors that
arrise when flushing (syncing) block data to disk.

This commit is part of the libbitcoinkernel project and further removes
the shutdown's and, more generally, the kernel library's dependency on
interface_ui with a kernel notification method. By removing interface_ui
from the kernel library, its dependency on boost is reduced to just
boost::multi_index. At the same time it also takes a step towards
de-globalising the interrupt infrastructure.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/27861 | core#27861]]
https://github.com/bitcoin/bitcoin/pull/27861/commits/7320db96f8d2aeff0bc5dc67d8b7b37f5f808990
Depends on D18967

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18968